fix: add exist check for vehicle destroying

+++ src/server/classes/property.lua
@@ -295,7 +295,9 @@ end
 ---Destroy all the vehicles in the property
 function Property:destroyVehicles()
     for _, vehicle in pairs(self.vehicles) do
-        DeleteEntity(vehicle.entity)
+        if DoesEntityExist(vehicle.entity) then
+            DeleteEntity(vehicle.entity)
+        end
         vehicle.entity = nil
     end
 end
